# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for alcotracker backend",
        "title": "Alcotracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Backend Dev",
            "email": "tommanger55@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/userinfo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve the profile of the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Userinfo"
                ],
                "summary": "Get User Profile",
                "operationId": "GetUserProfile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Server or Database internal error",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    }
                }
            }
        },
        "/userinfo/change-username": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the username of current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Userinfo"
                ],
                "summary": "Change Username",
                "operationId": "ChangeUsernamce",
                "parameters": [
                    {
                        "description": "New username info",
                        "name": "UsernameChangeRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userinfo.UsernameChangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Server or Database internal error",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    }
                }
            }
        },
        "/userinfo/follow": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Follow a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Userinfo"
                ],
                "summary": "Follow User",
                "operationId": "FollowUser",
                "parameters": [
                    {
                        "description": "Info about user to follow",
                        "name": "FollowRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userinfo.FollowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Server or Database internal error",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    }
                }
            }
        },
        "/userinfo/unfollow": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Unfollow a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Userinfo"
                ],
                "summary": "Unfollow User",
                "operationId": "UnfollowUser",
                "parameters": [
                    {
                        "description": "Info about user to unfollow",
                        "name": "FollowRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userinfo.FollowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Server or Database internal error",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    }
                }
            }
        },
        "/userinfo/userprofile": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the User with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Userinfo"
                ],
                "summary": "Get User By Id",
                "operationId": "GetUserById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user to retrieve",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Server or Database internal error",
                        "schema": {
                            "$ref": "#/definitions/GenericResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "GenericResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "userinfo.FollowRequest": {
            "type": "object",
            "properties": {
                "to_follow_id": {
                    "type": "string"
                }
            }
        },
        "userinfo.UsernameChangeRequest": {
            "type": "object",
            "properties": {
                "new_username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}